#117 PosterService -> PosterService, PosterServiceImpl
